added dtype to kwargs

+++ hdf5utils.py
@@ -122,7 +122,7 @@ class HDF5Handler(object):
             self.create_dset(dset_path, array, **kwargs) 
             self.index[dset_path].append(ndarray)
 
-    def create_dset(self, dset_path, array, chunksize=1000, blockfactor=100):
+    def create_dset(self, dset_path, array, chunksize=1000, blockfactor=100, dtype='float64'):
         """
         Define h5py dataset parameters here. 
 
@@ -155,7 +155,7 @@ class HDF5Handler(object):
         chunkshape = sum(((chunksize,), arr_shape), ())
         maxshape = sum(((None,), arr_shape), ())
 
-        dsetkw = dict(chunks=chunkshape, maxshape=maxshape)
+        dsetkw = dict(chunks=chunkshape, maxshape=maxshape, dtype=dtype)
                                        
         init_shape = sum(((blocksize,), arr_shape), ())
         dset = self.file.create_dataset(dset_path, shape=init_shape, **dsetkw)
@@ -192,7 +192,7 @@ class HDF5HandlerAmuse(HDF5Handler):
             self.create_dset(dset_path, array, **kwargs) 
             self.index[dset_path].append(ndarray)
 
-    def create_dset(self, dset_path, array, chunksize=1000, blockfactor=100):
+    def create_dset(self, dset_path, array, chunksize=1000, blockfactor=100, dtype='float64'):
         """
         Define h5py dataset parameters here. 
 
@@ -238,7 +238,7 @@ class HDF5HandlerAmuse(HDF5Handler):
         chunkshape = sum(((chunksize,), arr_shape), ())
         maxshape = sum(((None,), arr_shape), ())
 
-        dsetkw = dict(chunks=chunkshape, maxshape=maxshape)
+        dsetkw = dict(chunks=chunkshape, maxshape=maxshape, dtype=dtype)
         init_shape = sum(((blocksize,), arr_shape), ())
 
         dset = self.file.create_dataset(dset_path, shape=init_shape, **dsetkw)
